refactor: user to base api
