+++ backend/apps/api/serializers/user/__init__.py
@@ -1,0 +1,15 @@
+from rest_framework import serializers
+
+from apps.user.models import User
+
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = ('id', 'email', 'password', 'date_joined')
+        read_only_fields = ('date_joined',)
+        extra_kwargs = {'password': {'write_only': True}}
+
+    def create(self, validated_data):
+        user = User.objects.create_user(**validated_data)  # pyright: ignore
+        return user
