+++ backend/apps/api/views/user/auth.py
@@ -1,12 +1,12 @@
-from common.api.exceptions import ServerError
-from common.api.serializers import DetailResponseSerializer
 from django.contrib.auth import authenticate
 from drf_spectacular.utils import extend_schema
 from rest_framework import exceptions, generics, permissions, views
 from rest_framework.authtoken.models import Token
 from rest_framework.response import Response
 
-from .serializers import AuthSerializer, AuthTokenSerializer, ProfileSerializer
+from ...bases.serializers import DetailResponseSerializer
+from ...exceptions import ServerError
+from ...serializers.user.auth import AuthSerializer, AuthTokenSerializer
 
 
 class LoginAPIView(views.APIView):
@@ -54,26 +54,3 @@ class RegisterAPIView(generics.CreateAPIView):
     """
 
     serializer_class = AuthSerializer
-
-
-class MeAPIView(views.APIView):
-    """
-    View to retrieve information for the currently authenticated user.
-
-    - `get`: Returns the details of the authenticated user based on their token.
-
-    Permission:
-    - Requires user authentication.
-    """
-
-    permission_classes = (permissions.IsAuthenticated,)
-    serializer_class = ProfileSerializer
-
-    def get(self, request):
-        if request.user_profile:
-            serializer = self.serializer_class(
-                request.user_profile, context={'request': request}
-            )
-            return Response(serializer.data)
-        else:
-            raise exceptions.ValidationError('A valid profile must be provided.')
